Switched from textile to haml in basic flow section because I wanted a second paragraph within a list.
Finished writing the interface suggestion for when the supervisor
makes selection of category etc.

Added a wrongphase class that marks comments about the wrong phase of
the engineering.

Changed the table formatting to be more device independent (everything
was specified in px because I modified the css from someone else's web
page). Now font sizes are specified relative to the default and
margins etc are specifed in em's so everything should scale with the
default font.