Finessed the generated tests a bit:
Removed a bunch of redundant tests generated because pythoscope
couldn't get the constructors right.

Removed assert False statements where the generator expected you to
fill in tests and replaced them with expected failures.

The result is a mostly useless test suite, but one where all the
boiler-plate is done for me and I can focus on writing simple tests.